Refactor project structure and add media folder
