+++ core/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class CoreAppConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'core_app'
+    name = 'core'
