+++ core/models.py
@@ -1,5 +1,20 @@
 from django.db import models
-from puplic_app.models import User
+
+
+class User(models.Model):
+    username = models.CharField(max_length=150, unique=True)
+    email = models.EmailField(unique=True)
+    password = models.CharField(max_length=128)  
+    phone = models.CharField(max_length=20, blank=True, null=True)
+    role = models.CharField(
+        max_length=20,
+        choices=[('owner', 'Owner'), ('renter', 'Renter')],
+        default='renter'
+    )
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return self.username
 
 class Car(models.Model):
     owner = models.ForeignKey(User, on_delete=models.CASCADE, related_name='cars')
