--- puplic_app/urls.py
@@ -1,12 +1,0 @@
-from django.urls import path
-from . import views
-
-app_name = 'puplic_app'
-
-urlpatterns = [
-    path('', views.home, name='home'),             
-    path('about/', views.about, name='about'),     
-    path('contact/', views.contact, name='contact'),
-    path('login/', views.login, name='login'), 
-    path('register/', views.register, name='register'), 
-]
