+++ royalCarRental_project/settings.py
@@ -37,8 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'puplic_app',
-    'core_app',
+    'core',
 
 ]
